Splintering tests into own classes
